renamed Styles for aboutUs folder.
